project/requirements.txt
agate==1.9.1
annotated-types==0.7.0
attrs==25.1.0
azure-core==1.32.0
azure-eventhub==5.14.0
azure-storage-blob==12.24.1
babel==2.17.0
backports.tarfile==1.2.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
daff==1.3.46
databricks-cli==0.18.0
databricks-connect==15.4.5
databricks-dlt==0.3.0
databricks-sdk==0.29.0
databricks-sql-connector==3.7.2
dbt-adapters==1.14.0
dbt-common==1.15.0
dbt-core==1.9.2
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.7.4
dbt-spark==1.9.1
deepdiff==7.0.1
et_xmlfile==2.0.0
google-analytics-data==0.18.17
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.70.0
idna==3.10
importlib-metadata==6.11.0
iniconfig==2.0.0
isodate==0.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
leather==0.4.0
lz4==4.4.3
MarkupSafe==3.0.2
mashumaro==3.14
more-itertools==10.6.0
msgpack==1.1.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.5
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
parsedatetime==2.6
pathspec==0.12.1
pluggy==1.5.0
proto-plus==1.26.0
protobuf==5.29.3
py4j==0.10.9.7
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
rsa==4.9
six==1.17.0
snowplow-tracker==1.0.4
sqlparams==6.2.0
sqlparse==0.5.3
tabulate==0.9.0
text-unidecode==1.3
thrift==0.20.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zipp==3.21.0
